Move social media stuff to separate component + tidy up
