+++ components/admin/listing-form/ListingForm.tsx
@@ -1,7 +1,7 @@
 'use client'
 import { useEffect, useMemo } from 'react'
 import { useForm, FormProvider } from 'react-hook-form'
-import { AiOutlineLoading, AiOutlinePlus } from 'react-icons/ai'
+import { AiOutlineLoading } from 'react-icons/ai'
 import ReactSelect from 'react-select'
 import type { Options } from 'react-select'
 import type { Category } from '@prisma/client'
@@ -15,7 +15,7 @@ import EditorField from './RichTextEditor'
 import { Button } from '@components/ui/button'
 import { Input } from '@components/ui/input'
 import { Checkbox } from '@components/ui/checkbox'
-import { socialMediaPlatforms } from '@helpers/socials'
+import SocialMedia from './SocialMedia'
 import {
   Form,
   FormField,
@@ -35,7 +35,6 @@ import {
   SelectTrigger,
   SelectValue,
 } from '@components/ui/select'
-import { FaShareAlt, FaTrash } from 'react-icons/fa'
 
 const Map = dynamic(() => import('./Map'), {
   ssr: false,
@@ -223,16 +222,6 @@ const ListingForm = ({
       category: Number(submittedData.category),
     }
 
-    // Extract individual social media fields from the socialMedia array
-    if (data.socials && Array.isArray(data.socials)) {
-      // Convert the socialMedia array to individual fields for backward compatibility
-      data.socials.forEach((item) => {
-        if (item.platform && item.url) {
-          data[item.platform] = item.url
-        }
-      })
-    }
-
     const isNewImage = data.image instanceof File
     if (!isNewImage && data.image) {
       delete data.image
@@ -378,209 +367,7 @@ const ListingForm = ({
             )}
           />
 
-          <div className="mt-6">
-            <div className="flex items-center justify-between pb-2">
-              <FormLabel className="font-semibold">Social Media</FormLabel>
-              <Button
-                type="button"
-                variant="outline"
-                size="sm"
-                className="flex items-center gap-1"
-                onClick={() => {
-                  const currentSocials = methods.getValues('socials') || []
-                  // Find a platform that hasn't been used yet
-                  const unusedPlatforms = socialMediaPlatforms.filter(
-                    (platform) =>
-                      !currentSocials.some(
-                        (item) => item.platform === platform.id,
-                      ),
-                  )
-
-                  if (unusedPlatforms.length > 0) {
-                    methods.setValue('socials', [
-                      ...currentSocials,
-                      { platform: unusedPlatforms[0].id, url: '' },
-                    ])
-                  }
-                }}
-                disabled={
-                  methods.getValues('socials')?.length >=
-                  socialMediaPlatforms.length
-                }
-              >
-                <AiOutlinePlus className="mr-1" />
-                Add Social Media
-              </Button>
-            </div>
-
-            <div className="flex flex-col gap-1">
-              {methods.watch('socials')?.map((item, index) => {
-                const platform = socialMediaPlatforms.find(
-                  (p) => p.id === item.platform,
-                )
-                return (
-                  <div
-                    key={index}
-                    className="group relative rounded-md border border-gray-200 bg-white p-4 shadow-sm transition-all hover:shadow-md"
-                  >
-                    <div className="flex items-end gap-4">
-                      <div className="w-1/3">
-                        <FormField
-                          control={methods.control}
-                          name={`socials.${index}.platform`}
-                          render={({ field }) => (
-                            <FormItem className="space-y-1">
-                              <FormLabel className="text-sm text-gray-600">
-                                Platform
-                              </FormLabel>
-                              <Select
-                                value={field.value}
-                                onValueChange={(value) => {
-                                  field.onChange(value)
-                                }}
-                              >
-                                <SelectTrigger className="bg-white">
-                                  <SelectValue placeholder="Select platform">
-                                    {field.value && (
-                                      <div className="flex items-center gap-2">
-                                        {(() => {
-                                          const platform =
-                                            socialMediaPlatforms.find(
-                                              (p) => p.id === field.value,
-                                            )
-                                          if (platform) {
-                                            const Icon = platform.icon
-                                            return (
-                                              <Icon
-                                                style={{
-                                                  color: platform.color,
-                                                }}
-                                              />
-                                            )
-                                          }
-                                          return null
-                                        })()}
-                                        {
-                                          socialMediaPlatforms.find(
-                                            (p) => p.id === field.value,
-                                          )?.label
-                                        }
-                                      </div>
-                                    )}
-                                  </SelectValue>
-                                </SelectTrigger>
-                                <SelectContent>
-                                  {socialMediaPlatforms
-                                    .filter((platform) => {
-                                      // Allow the current platform or platforms not already selected
-                                      const currentPlatforms = methods
-                                        .getValues('socials')
-                                        .map((item) => item.platform)
-                                      return (
-                                        platform.id === field.value ||
-                                        !currentPlatforms.includes(platform.id)
-                                      )
-                                    })
-                                    .map((platform) => (
-                                      <SelectItem
-                                        key={platform.id}
-                                        value={platform.id}
-                                      >
-                                        <div className="flex items-center gap-2">
-                                          {(() => {
-                                            const Icon = platform.icon
-                                            return (
-                                              <Icon
-                                                style={{
-                                                  color: platform.color,
-                                                }}
-                                              />
-                                            )
-                                          })()}
-                                          {platform.label}
-                                        </div>
-                                      </SelectItem>
-                                    ))}
-                                </SelectContent>
-                              </Select>
-                              <FormMessage />
-                            </FormItem>
-                          )}
-                        />
-                      </div>
-
-                      <div className="flex-1">
-                        <FormField
-                          control={methods.control}
-                          name={`socials.${index}.url`}
-                          render={({ field }) => (
-                            <FormItem className="space-y-1">
-                              <FormLabel className="text-sm text-gray-600">
-                                URL
-                              </FormLabel>
-                              <div className="relative">
-                                <FormControl>
-                                  <Input
-                                    {...field}
-                                    placeholder={platform?.prefix || 'https://'}
-                                  />
-                                </FormControl>
-                              </div>
-                              <FormMessage />
-                            </FormItem>
-                          )}
-                        />
-                      </div>
-
-                      <div className="ml-2 flex-shrink-0 self-end">
-                        <Button
-                          type="button"
-                          variant="ghost"
-                          size="icon"
-                          className="h-10 w-10 rounded-full opacity-70 hover:bg-red-50 hover:text-red-600 hover:opacity-100"
-                          onClick={() => {
-                            const currentSocials = [
-                              ...methods.getValues('socials'),
-                            ]
-                            currentSocials.splice(index, 1)
-                            methods.setValue('socials', currentSocials)
-                          }}
-                        >
-                          <FaTrash />
-                        </Button>
-                      </div>
-                    </div>
-                  </div>
-                )
-              })}
-
-              {(!methods.watch('socials') ||
-                methods.watch('socials').length === 0) && (
-                <div className="flex flex-col items-center justify-center rounded-md border border-dashed border-gray-300 bg-gray-50 p-8 text-center">
-                  <div className="mb-2 rounded-full bg-gray-100 p-3">
-                    <FaShareAlt className="text-2xl text-gray-400" />
-                  </div>
-                  <p className="text-muted-foreground mb-2">
-                    No social media links added
-                  </p>
-                  <Button
-                    type="button"
-                    variant="outline"
-                    size="sm"
-                    className="mt-2"
-                    onClick={() => {
-                      methods.setValue('socials', [
-                        { platform: socialMediaPlatforms[0].id, url: '' },
-                      ])
-                    }}
-                  >
-                    <AiOutlinePlus className="mr-1" />
-                    Add Social Media
-                  </Button>
-                </div>
-              )}
-            </div>
-          </div>
+          <SocialMedia />
 
           <div className="mt-4">
             <SlugField
