+++ components/main-list/dialog/Dialog.tsx
@@ -4,7 +4,6 @@ import { memo, useCallback, useEffect, useMemo, useState } from 'react'
 import Link from 'next/link'
 import Image from 'next/image'
 import { useSearchParams } from 'next/navigation'
-import { FaFacebook, FaTwitter, FaInstagram } from 'react-icons/fa'
 import { HiUserGroup, HiOutlineLink, HiExternalLink } from 'react-icons/hi'
 import { toast } from 'sonner'
 import { sanitizeLink } from '@helpers/utils'
